requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
croniter==6.0.0
Django==6.0
django-debug-toolbar==6.1.0
django-redis==6.0.0
django-rq==3.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
rq==2.6.1
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0
